gym-oaas/requirements.txt
absl-py==0.13.0
anyio==3.2.1
argon2-cffi==20.1.0        
async-generator==1.10      
atari-py==0.2.9
atomicwrites==1.4.0        
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cloudpickle==1.6.0
coverage==5.3
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
google-auth==1.32.1        
google-auth-oauthlib==0.4.4
grpcio==1.38.1
gym==0.18.3
-e git+https://github.com/ArnaudFickinger/gym-multigrid.git@77b0ee0053e1c6477328a7d862cb326a6bc4cd6a#egg=gym_multigrid
# Editable install with no version control (gym-oaas==0.0.1)
-e /local_data/gym-oaas
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==4.6.1
intervaltree==3.0.2
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
kiwisolver==1.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.0
matplotlib-inline==0.1.2
-e git+https://github.com/zuoxingdong/mazelab.git@236eef5f7c41b86bb784f506fe1a2e0700a2e48f#egg=mazelab
mistune==0.8.4
more-itertools==8.8.0
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.1
notebook==6.4.0
numpy==1.21.0
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.2.54
packaging==21.0
pandas==1.3.0
pandocfilters==1.4.3
parallelworkloads==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pip-install-test==0.5
pluggy==0.13.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.15
Pygments==2.9.0
pyparsing==2.4.6
pyrsistent==0.18.0
pytest==4.6.3
pytest-cov==2.12.1
python-dateutil==2.7.3
pytz==2021.1
PyWavelets==1.1.1
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-image==0.18.2
scipy==1.7.0
Send2Trash==1.7.1
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
stable-baselines3==1.1.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.10.1
testpath==0.5.0
tifffile==2021.7.2
toml==0.10.2
torch==1.9.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
zipp==3.5.0